move composition to submodules
